practice/chapter9/code/requirements.txt
# 9장 실습: GraphRAG/LightRAG 비교
# 설치: pip install -r requirements.txt

langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
chromadb>=0.5.0
openai>=1.50.0
python-dotenv>=1.0.0
networkx>=3.0
